python/surrealfs_ai/pyproject.toml
[project]
name = "surrealfs-ai"
version = "0.1.0"
dependencies = [
    "surrealfs-py",
    "pydantic>=2.12.5",
    "pydantic-ai-slim[web]>=1.47.0",
    "openai==2.16.0",
]

[project.optional-dependencies]
demo = [
    "logfire>=4.20.0",
    "pydantic-ai-slim[web,anthropic]>=1.47.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
surrealfs-py = { workspace = true }
